*: Enable GPG support for metadata commands
