+++ internal/cmd/policy/init/init.go
@@ -6,7 +6,6 @@ package init
 import (
 	"github.com/gittuf/gittuf/experimental/gittuf"
 	trustpolicyopts "github.com/gittuf/gittuf/experimental/gittuf/options/trustpolicy"
-	"github.com/gittuf/gittuf/internal/cmd/common"
 	"github.com/gittuf/gittuf/internal/cmd/policy/persistent"
 	"github.com/gittuf/gittuf/internal/policy"
 	"github.com/spf13/cobra"
@@ -50,7 +49,6 @@ func New(persistent *persistent.Options) *cobra.Command {
 		Use:               "init",
 		Short:             "Initialize policy file",
 		Long:              "This command initializes a new gittuf policy file with the specified policy name. The user must provide a valid signing key.",
-		PreRunE:           common.CheckForSigningKeyFlag,
 		RunE:              o.Run,
 		DisableAutoGenTag: true,
 	}
