+++ internal/cmd/trust/disablegithubappapprovals/disablegithubappapprovals.go
@@ -6,7 +6,6 @@ package disablegithubappapprovals
 import (
 	"github.com/gittuf/gittuf/experimental/gittuf"
 	trustpolicyopts "github.com/gittuf/gittuf/experimental/gittuf/options/trustpolicy"
-	"github.com/gittuf/gittuf/internal/cmd/common"
 	"github.com/gittuf/gittuf/internal/cmd/trust/persistent"
 	"github.com/gittuf/gittuf/internal/tuf"
 	"github.com/spf13/cobra"
@@ -50,7 +49,6 @@ func New(persistent *persistent.Options) *cobra.Command {
 		Use:               "disable-github-app-approvals",
 		Short:             "Mark GitHub app approvals as untrusted henceforth",
 		Long:              `The 'disable-github-app-approvals' command revokes a GitHub App's ability to approve changes by marking it untrusted in the trust policy.`,
-		PreRunE:           common.CheckForSigningKeyFlag,
 		RunE:              o.Run,
 		DisableAutoGenTag: true,
 	}
