+++ internal/cmd/trust/updatepolicythreshold/updatepolicythreshold.go
@@ -6,7 +6,6 @@ package updatepolicythreshold
 import (
 	"github.com/gittuf/gittuf/experimental/gittuf"
 	trustpolicyopts "github.com/gittuf/gittuf/experimental/gittuf/options/trustpolicy"
-	"github.com/gittuf/gittuf/internal/cmd/common"
 	"github.com/gittuf/gittuf/internal/cmd/trust/persistent"
 	"github.com/spf13/cobra"
 )
@@ -47,11 +46,10 @@ func (o *options) Run(cmd *cobra.Command, _ []string) error {
 func New(persistent *persistent.Options) *cobra.Command {
 	o := &options{p: persistent}
 	cmd := &cobra.Command{
-		Use:     "update-policy-threshold",
-		Short:   "Update Policy threshold in the gittuf root of trust",
-		Long:    "This command allows users to update the threshold of valid signatures required for the policy.",
-		PreRunE: common.CheckForSigningKeyFlag,
-		RunE:    o.Run,
+		Use:   "update-policy-threshold",
+		Short: "Update Policy threshold in the gittuf root of trust",
+		Long:  "This command allows users to update the threshold of valid signatures required for the policy.",
+		RunE:  o.Run,
 	}
 	o.AddFlags(cmd)
